Mock tasks moved to engine
