+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/mock/MockBackgroundTask.java
@@ -14,27 +14,67 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
- *
  */
 
-package ai.grakn.test.engine.tasks;
+package ai.grakn.engine.tasks.mock;
 
 import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.BackgroundTask;
+import com.google.common.collect.ConcurrentHashMultiset;
+import com.google.common.collect.ImmutableMultiset;
 import mjson.Json;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
 
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.addCancelledTask;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.addCompletedTask;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.onTaskFinish;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.onTaskStart;
-
 public abstract class MockBackgroundTask implements BackgroundTask {
+
+    private static final ConcurrentHashMultiset<TaskId> COMPLETED_TASKS = ConcurrentHashMultiset.create();
+    private static final ConcurrentHashMultiset<TaskId> CANCELLED_TASKS = ConcurrentHashMultiset.create();
+    private static Consumer<TaskId> onTaskStart;
+    private static Consumer<TaskId> onTaskFinish;
 
     protected final AtomicBoolean cancelled = new AtomicBoolean(false);
     protected final Object sync = new Object();
+
+    static void addCompletedTask(TaskId taskId) {
+        COMPLETED_TASKS.add(taskId);
+    }
+
+    public static ImmutableMultiset<TaskId> completedTasks() {
+        return ImmutableMultiset.copyOf(COMPLETED_TASKS);
+    }
+
+    static void addCancelledTask(TaskId taskId) {
+        CANCELLED_TASKS.add(taskId);
+    }
+
+    public static ImmutableMultiset<TaskId> cancelledTasks() {
+        return ImmutableMultiset.copyOf(CANCELLED_TASKS);
+    }
+
+    public static void whenTaskStarts(Consumer<TaskId> beforeTaskStarts) {
+        MockBackgroundTask.onTaskStart = beforeTaskStarts;
+    }
+
+    static void onTaskStart(TaskId taskId) {
+        if (onTaskStart != null) onTaskStart.accept(taskId);
+    }
+
+    public static void whenTaskFinishes(Consumer<TaskId> onTaskFinish) {
+        MockBackgroundTask.onTaskFinish = onTaskFinish;
+    }
+
+    static void onTaskFinish(TaskId taskId) {
+        if (onTaskFinish != null) onTaskFinish.accept(taskId);
+    }
+
+    public static void clearTasks() {
+        COMPLETED_TASKS.clear();
+        CANCELLED_TASKS.clear();
+        onTaskStart = null;
+        onTaskFinish = null;
+    }
 
     @Override
     public final boolean start(Consumer<String> saveCheckpoint, Json configuration) {
