+++ grakn-test/src/test/java/ai/grakn/test/engine/controller/TasksControllerTest.java
@@ -23,8 +23,8 @@ import ai.grakn.engine.tasks.TaskManager;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.test.EngineContext;
 import ai.grakn.test.engine.tasks.BackgroundTaskTestUtils;
-import ai.grakn.test.engine.tasks.LongExecutionTestTask;
-import ai.grakn.test.engine.tasks.ShortExecutionTestTask;
+import ai.grakn.engine.tasks.mock.LongExecutionTestTask;
+import ai.grakn.engine.tasks.mock.ShortExecutionTestTask;
 import com.jayway.restassured.http.ContentType;
 import com.jayway.restassured.response.Response;
 import org.json.JSONArray;
