+++ grakn-test/src/test/java/ai/grakn/test/engine/tasks/manager/singlequeue/SingleQueueTaskRunnerTest.java
@@ -25,9 +25,9 @@ import ai.grakn.engine.tasks.manager.singlequeue.SingleQueueTaskManager;
 import ai.grakn.engine.tasks.manager.singlequeue.SingleQueueTaskRunner;
 import ai.grakn.engine.tasks.storage.TaskStateInMemoryStore;
 import ai.grakn.engine.util.EngineID;
-import ai.grakn.test.engine.tasks.EndlessExecutionTestTask;
-import ai.grakn.test.engine.tasks.LongExecutionTestTask;
-import ai.grakn.test.engine.tasks.ShortExecutionTestTask;
+import ai.grakn.engine.tasks.mock.EndlessExecutionTestTask;
+import ai.grakn.engine.tasks.mock.LongExecutionTestTask;
+import ai.grakn.engine.tasks.mock.ShortExecutionTestTask;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMultiset;
@@ -59,14 +59,14 @@ import static ai.grakn.engine.TaskStatus.STOPPED;
 import static ai.grakn.engine.tasks.TaskSchedule.at;
 import static ai.grakn.engine.tasks.TaskSchedule.recurring;
 import static ai.grakn.test.GraknTestEnv.hideLogs;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.cancelledTasks;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.clearTasks;
+import static ai.grakn.engine.tasks.mock.MockBackgroundTask.cancelledTasks;
+import static ai.grakn.engine.tasks.mock.MockBackgroundTask.clearTasks;
 import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.completableTasks;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.completedTasks;
+import static ai.grakn.engine.tasks.mock.MockBackgroundTask.completedTasks;
 import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.createTask;
 import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.failingTasks;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.whenTaskFinishes;
-import static ai.grakn.test.engine.tasks.BackgroundTaskTestUtils.whenTaskStarts;
+import static ai.grakn.engine.tasks.mock.MockBackgroundTask.whenTaskFinishes;
+import static ai.grakn.engine.tasks.mock.MockBackgroundTask.whenTaskStarts;
 import static java.time.Duration.between;
 import static java.time.Duration.ofMillis;
 import static java.time.Instant.now;
