+++ grakn-test/src/test/java/ai/grakn/test/engine/tasks/storage/TaskStateGraphStoreTest.java
@@ -25,7 +25,7 @@ import ai.grakn.engine.tasks.TaskStateStorage;
 import ai.grakn.engine.tasks.storage.TaskStateGraphStore;
 import ai.grakn.engine.util.EngineID;
 import ai.grakn.test.EngineContext;
-import ai.grakn.test.engine.tasks.ShortExecutionTestTask;
+import ai.grakn.engine.tasks.mock.ShortExecutionTestTask;
 import mjson.Json;
 import org.junit.Assert;
 import org.junit.Before;
